requirements.txt
APScheduler==3.11.2
asgiref==3.11.0
attrs==25.4.0
certifi==2026.1.4
charset-normalizer==3.4.4
Django==5.2.10
django-cors-headers==4.9.0
django-filter==25.2
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.1
drf-spectacular==0.29.0
drf-yasg==1.21.11
Faker==40.1.0
greenlet==3.3.0
idna==3.11
inflection==0.5.1
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
mysqlclient==2.2.7
packaging==25.0
playwright==1.57.0
pyee==13.0.0
PyJWT==2.10.1
python-dotenv==1.2.1
pytz==2025.2
PyYAML==6.0.3
referencing==0.37.0
requests==2.32.5
rpds-py==0.30.0
sqlparse==0.5.5
typing_extensions==4.15.0
tzdata==2025.3
tzlocal==5.3.1
uritemplate==4.2.0
urllib3==2.6.3
python-dotenv==1.2.1
pytz==2025.2
beautifulsoup4==4.14.3